Create NestJs Dynamic Module for Airtable
This allows an Airtable instance to be configured within a Nest dynamic module.

This makes the use of Airtable more seamless within a Nestjs application. The dynamic module exposes the usual `forRoot()` and `forRootAsync()` static methods that allow the module to be configured with the Airtable api key and any properties determined by the AirtableJs library.